+++ src/java/abfab3d/grid/ArrayAttributeGridShort.java
@@ -38,8 +38,8 @@ public class ArrayAttributeGridShort extends BaseAttributeGrid {
      */
     public ArrayAttributeGridShort(double w, double h, double d, double pixel, double sheight) {
         this((int) (Math.ceil(w / pixel)) + 1,
-        	 (int) (Math.ceil(h / sheight)) + 1,
-             (int) (Math.ceil(d / pixel)) + 1, 
+             (int) (Math.ceil(h / sheight)) + 1,
+             (int) (Math.ceil(d / pixel)) + 1,
              pixel,
              sheight);
     }
@@ -178,7 +178,7 @@ public class ArrayAttributeGridShort extends BaseAttributeGrid {
      * @param y The y world coordinate
      * @param z The z world coordinate
      */
-    public int getAttribute(double x, double y, double z) {
+    public long getAttribute(double x, double y, double z) {
         int slice = (int) (y / sheight);
         int s_x = (int) (x / pixelSize);
         int s_z = (int) (z / pixelSize);
@@ -197,7 +197,7 @@ public class ArrayAttributeGridShort extends BaseAttributeGrid {
      * @param y The y world coordinate
      * @param z The z world coordinate
      */
-    public int getAttribute(int x, int y, int z) {
+    public long getAttribute(int x, int y, int z) {
         int idx = y * sliceSize + x * depth + z;
 
         short mat = (short) (0x3FFF & data[idx]);
@@ -214,7 +214,7 @@ public class ArrayAttributeGridShort extends BaseAttributeGrid {
      * @param state The voxel state
      * @param material The material
      */
-    public void setData(double x, double y, double z, byte state, int material) {
+    public void setData(double x, double y, double z, byte state, long material) {
         int slice = (int) (y / sheight);
         int s_x = (int) (x / pixelSize);
         int s_z = (int) (z / pixelSize);
@@ -233,7 +233,7 @@ public class ArrayAttributeGridShort extends BaseAttributeGrid {
      * @param state The voxel state
      * @param material The material
      */
-    public void setData(int x, int y, int z, byte state, int material) {
+    public void setData(int x, int y, int z, byte state, long material) {
         int idx = y * sliceSize + x * depth + z;
 
         data[idx] = (short) (0xFFFF & (((short)state) << 14 | (short)material));
@@ -247,7 +247,7 @@ public class ArrayAttributeGridShort extends BaseAttributeGrid {
      * @param z The z world coordinate
      * @param material The materialID
      */
-    public void setAttribute(int x, int y, int z, int material) {
+    public void setAttribute(int x, int y, int z, long material) {
         int idx = y * sliceSize + x * depth + z;
 
         byte state = (byte) ((data[idx] & 0xFFFF) >> 14);
